dev (#86)
* added rt alias for sudo -i and edited help message. will soon try to come up with a way to automatically update the help message in custom_functions.zsh

* Update with some slight changes and tweaking, mainly due to brew and gh completions.

* Moved some things from .zshenv to .zshrc

* Added my own p10k-promptconfig plugin to be installed in install.sh, modified some other things and removed the P10K-themes directory since it's no longer needed.

* QoL updates to install.sh, now checks if zsh is already default shell (using $SHELL environment variable) before attempting to change it

* Removed some unnecessary lines from .zshrc and .zshenv; removed .zprofile

* Shortened some aliases

* updated .gitignore

* added function to install.sh for plugin (and theme) cloning/updating, plus other minor updates

+++ omz-files/custom_aliases.zsh
@@ -1,7 +1,7 @@
 # ALIASES
 
 # alias should only be set if on Parrot OS
-if var=$(cat /etc/os-release | grep -wo 'parrot'); then
+if (( $+commands[parrot-upgrade] )); then
     alias sys-update='sudo parrot-upgrade'
 fi
 
@@ -10,11 +10,8 @@ alias s='sudo'
 alias rt='sudo -i'
 alias e='echo'
 alias update-db="sudo updatedb --prunepaths='/timeshift/snapshots /media /run/timeshift /run/user'"
-alias sysinfo='neofetch && info-message'
-alias fk='fuck'
 alias own='sudo chown -v $USER:$USER'
 alias owndir='sudo chown -R $USER:$USER'
-alias c='conda activate'
 alias z='cd $ZDOTDIR'
 alias zdir="$ZDOTDIR/"
 # copy, symlinks
@@ -25,36 +22,55 @@ alias sl-ffs='sudo ln -srv'
 # clean zcomp and zwc zsh files
 alias cleanzsh='rm -rf ${ZDOTDIR:-$HOME}/.zcompdump* && rm -rf ${ZDOTDIR:-$HOME}/.zshrc.zwc'
 
-# apt/apt-get/dpkg
-alias afix='sudo apt-get install -f'
-alias alu='apt list --upgradeable'
-alias au='sudo apt-get update'
-alias ai='sudo apt install'
-alias ar='sudo apt reinstall'
-alias arm='sudo apt remove'
-alias ap='sudo apt purge'
-alias af='sudo apt search'
-alias ash='sudo apt show'
-alias ag='sudo apt-get'
-alias a='sudo apt'
-alias ac='sudo apt-cache'
-alias mark-a='sudo apt-mark auto'
-alias mark-m='sudo apt-mark manual'
-alias aptc='sudo apt-get --purge autoremove -y && sudo apt-get autoclean -y'
+if (( $+commands[neofetch] )); then
+	alias sinfo='neofetch && info-message'
+fi
+
+if (( $+commands[fuck] )); then
+alias fk='fuck'
+fi
+
+# apt/apt-get
+if (( $+commands[apt] && $+commands[apt-get] )); then
+	alias a='sudo apt'
+	alias ag='sudo apt-get'
+	alias ac='sudo apt-cache'
+	alias alu='apt list --upgradeable'
+	alias ase='apt search'
+	alias ash='apt show'
+	alias au='sudo apt-get update'
+	alias ai='sudo apt install'
+	alias ar='sudo apt reinstall'
+	alias arm='sudo apt remove'
+	alias ap='sudo apt purge'
+	alias mark-a='sudo apt-mark auto'
+	alias mark-m='sudo apt-mark manual'
+	alias afix='sudo apt-get install -f'
+	alias apc='sudo apt-get --purge autoremove -y && sudo apt-get autoclean -y'
+fi
+
 alias reconf='sudo dpkg-reconfigure'
 alias add-arch='sudo dpkg --add-architecture'
 
 # git aliases
-alias prcreate='gh pr create --fill'
-alias prmerge='gh pr merge'
+if (( $+commands[gh] )); then
+	alias prcreate='gh pr create --fill'
+	alias prmerge='gh pr merge'
+fi
 
 # docker aliases
-alias d='docker'
-alias dps='docker ps'
+if (( $+commands[docker] )); then
+	alias d='docker'
+	alias dps='docker ps'
+fi
+
+if (( $+commands[conda] )); then
+	alias c='conda activate'
+fi
 
 # ip info aliases and other system aliases
 alias extip='curl https://ipecho.net/plain; echo'
 alias intip='hostname -I; echo'
 alias shutdown='sudo shutdown now'
-alias sysctl='sudo systemctl'
-alias sysd='sudo systemd'
+alias sctl='sudo systemctl'
+alias sd='sudo systemd'
